initial push of new ansible-dnac collection (#1)
* adds .gitignore to repo

* adds new httpapi connection plugin for dnac

This commit adds a new httpapi pluging that will communicate with the
Cisco Intent API provided by DNA Center.  The initial version of the
httpapi does not provide any configurable options beyond what is
supported by the httpapi connection plugin.

To use this plugin, set the value of ansible_network_os to the
full name of the plugin (ciscodevnet.ansible_dnac.dnac)

* adds new module devices

This commit adds a new module to the collection for managing devices in
inventory.  The devices inventory module can configure either a single
device in inventory or all devices.  This commit also adds the missing
COPYING license text file.

* adds new module facts

This commit adds a new module and module framework for collecting facts
from the DNAC server.  The commit also adds the initial fact collectors
for devices, interfaces and sites.

* adds new module api and new role dnac_api

This commit adds both a new module and a new role.  The new module is
api and can return the contents of an api call using the httpapi
connection plugin.  The new role provides a set of endpoint to call
using the api module